icu-service/pyproject.toml
[project]
name = "icu-service"
version = "0.1.0"
description = "ICU microservice for text normalization, transliteration, and transformation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "PyICU>=2.13",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "python-multipart>=0.0.9",
    # PySpark integration dependencies
    "pyspark>=3.4.0",
    "pandas>=2.0.0",
    "pyarrow>=13.0.0",
    "requests>=2.31.0",
    "tenacity>=8.2.3",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "basedpyright>=1.22.0",
    "pytest>=8.3.0",
    "httpx>=0.28.0",
]

[tool.ruff]
line-length = 140
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4"]
ignore = ["E501"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.basedpyright]
typeCheckingMode = "strict"
reportMissingModuleSource = false
reportMissingImports = false
exclude = ["__pycache__", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"
